+++ packages/react/src/components/Calendar/src/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./CustomCaptionLabel";
+export * from "./CustomDay";
+export * from "./Wrapper";
+export * from "./Weekdays";