+++ packages/react/src/components/Calendar/src/index.ts
@@ -1,0 +1,5 @@
+import { Calendar } from "./Calendar";
+
+export { Calendar } from "./Calendar";
+export type { CalendarProps } from "./calendar.types";
+export default Calendar;